requirements.txt
nltk
spacy
python-docx
scikit_learn
mysql-connector-python
streamlit